refactor(gfx): 移除废弃的 wgpu 键盘渲染器
删除未使用的 KeyboardRenderer 管线与 shader，统一黑键判断逻辑，仅保留 KeyInstance 类型及对应测试。

+++ crates/editor/gfx/src/miditrail_renderer/instances/tests.rs
@@ -36,7 +36,7 @@ fn test_key_positions() {
     let white_total: f32 = positions
         .iter()
         .enumerate()
-        .filter(|(i, _)| !is_black_key(*i as u32))
+        .filter(|(i, _)| !is_black_key(*i as isize))
         .map(|(i, _)| widths[i])
         .sum();
     assert!((white_total - 1.0).abs() < 1e-5);
